# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Muta — Offline Adaptive Tutor
  description: ADTC 2026. The only public surface is the versioned /v1 contract; every client — browser,
    phones, eval.py, CLI — speaks only /v1.
  version: 0.1.0
paths:
...
  /v1/audio/transcribe:
    post:
      tags:
      - audio
      summary: Audio Transcribe
      description: 'Uploaded audio → text. 503 when ASR is unavailable (the UI tells the student to

        type instead), 422 when ffmpeg can''t decode the file, 413 when the upload is too large.'
      operationId: audio_transcribe_v1_audio_transcribe_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_audio_transcribe_v1_audio_transcribe_post'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranscribeResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
...
    Body_audio_transcribe_v1_audio_transcribe_post:
      properties:
        audio:
          type: string
          contentMediaType: application/octet-stream
          title: Audio
        conversation_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Conversation Id
        student_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Student Id
      type: object
      required:
      - audio
      title: Body_audio_transcribe_v1_audio_transcribe_post
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    TranscribeResponse:
      properties:
        text:
          type: string
          title: Text
          description: What the ASR heard; empty when nothing was recognised.
        attachment_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Attachment Id
          description: Stored copy of the uploaded audio.
      type: object
      required:
      - text
      title: TranscribeResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...